+++ src/lib.rs
@@ -23,10 +23,10 @@ mod node;
 mod nonce;
 mod noop_stream;
 mod package_manifest;
-mod system;
 mod run;
 mod rustup;
 mod safe_encoding;
+mod system;
 mod toolchain;
 mod utils;
