+++ src/system/mod.rs
@@ -1,3 +1,3 @@
-pub (crate) mod date;
-pub (crate) mod rng;
-pub (crate) mod sleep;
+pub(crate) mod date;
+pub(crate) mod rng;
+pub(crate) mod sleep;
